pyproject.toml
[project]
name = "pfe-traj-opt"
version = "0.1.0"
description = "Code pour l'optimisation de trajectoires d'avion avec de l'intelligence artificielle"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gymnasium>=1.2.3",
    "openap>=2.4",
    "pygame-ce>=2.5.6",
    "rich>=14.2.0",
    "stable-baselines3>=2.7.1",
    "tensorboard>=2.20.0",
    "tqdm>=4.67.1",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["xpc*", "xp_sim_gym*"]
